+++ src/FileHelper/PathInfoInterface.php
@@ -14,6 +14,7 @@ namespace AppUtils\FileHelper;
 use AppUtils\FileHelper;
 use AppUtils\FileHelper_Exception;
 use AppUtils\Interface_Stringable;
+use SplFileInfo;
 
 /**
  * Interface for the file and folder info classes.
@@ -28,12 +29,42 @@ interface PathInfoInterface extends Interface_Stringable
     public function getExtension(bool $lowercase=true) : string;
     public function getPath() : string;
     public function getFolderPath() : string;
+
+    /**
+     * @return int The size on disk, in bytes.
+     */
+    public function getSize() : int;
     public function exists() : bool;
     public function isFolder() : bool;
     public function isFile() : bool;
     public function isIndeterminate() : bool;
     public function isWritable() : bool;
     public function isReadable() : bool;
+
+    /**
+     * Whether this file or folder's path is located within the target path:
+     * Returns true if the path equals to the target path, or is a subfolder
+     * of the target path.
+     *
+     * Examples, assuming this is a folder with the following path:
+     *
+     * <code>/home/user/someone</code>
+     *
+     * Checking target paths against this:
+     *
+     * - /home/user/someone = true (same path)
+     * - /home/user/someone/readme.txt = true (file within the path)
+     * - /home/user/someone/subfolder = true (subfolder of the path)
+     * - /path/to/folder = false (different path)
+     * - /home/user = false (parent folder of the path)
+     *
+     * NOTE: Both paths must exist on disk to resolve relative
+     * paths with <code>../</code> parts.
+     *
+     * @param string|PathInfoInterface|SplFileInfo $targetPath
+     * @return bool
+     */
+    public function isWithinPath($targetPath) : bool;
 
     /**
      * @return $this