+++ tests/testsuites/FileHelperTests/PathInfoTest.php
@@ -78,4 +78,85 @@ class PathInfoTest extends FileHelperTestCase
             FileHelper::getPathInfo('name.')
         );
     }
+
+    public function test_isWithinPath() : void
+    {
+        $baseFolder = $this->assetsFolder;
+
+        $this->assertFileExists($baseFolder);
+
+        $tests = array(
+            array(
+                'label' => 'File within folder',
+                'source' => $baseFolder.'/FolderTree/SubFolderA',
+                'target' => $baseFolder.'/FolderTree/SubFolderA/readme.md',
+                'expected' => true
+            ),
+            array(
+                'label' => 'File within subfolder',
+                'source' => $baseFolder.'/FolderTree/SubFolderB',
+                'target' => $baseFolder.'/FolderTree/SubFolderB/SubSubFolder/SubSubSubFolder/readme.md',
+                'expected' => true
+            ),
+            array(
+                'label' => 'Subfolder within folder',
+                'source' => $baseFolder.'/FolderTree/SubFolderB',
+                'target' => $baseFolder.'/FolderTree/SubFolderB/SubSubFolder',
+                'expected' => true
+            ),
+            array(
+                'label' => 'Parent folder',
+                'source' => $baseFolder.'/FolderTree/SubFolderB',
+                'target' => $baseFolder,
+                'expected' => false
+            ),
+            array(
+                'label' => 'Other folder',
+                'source' => $baseFolder,
+                'target' => __DIR__,
+                'expected' => false
+            ),
+            array(
+                'label' => 'File within file',
+                'source' => $baseFolder.'/copy-file.txt',
+                'target' => $baseFolder.'/42-bytes.txt',
+                'expected' => true
+            ),
+            array(
+                'label' => 'Subfolder file within file',
+                'source' => $baseFolder.'/42-bytes.txt',
+                'target' => $baseFolder.'/FolderTree/SubFolderA/readme.md',
+                'expected' => true
+            ),
+            array(
+                'label' => 'Subfolder within file',
+                'source' => $baseFolder.'/42-bytes.txt',
+                'target' => $baseFolder.'/FolderTree/',
+                'expected' => true
+            ),
+            array(
+                'label' => 'Parent folder within file',
+                'source' => $baseFolder.'/FolderTree/SubFolderA/readme.md',
+                'target' => $baseFolder.'/FolderTree',
+                'expected' => false
+            ),
+            array(
+                'label' => 'Relative path shenanigans',
+                'source' => $baseFolder.'/FolderTree/SubFolderA/readme.md',
+                'target' => $baseFolder.'/FolderTree/SubFolderB/SubSubFolder/../../SubFolderA/readme.md',
+                'expected' => true
+            )
+        );
+
+        foreach($tests as $test)
+        {
+            $this->assertSame(
+                $test['expected'],
+                FileHelper::getPathInfo($test['target'])->isWithinPath($test['source']),
+                'Checked test "'.$test['label'].'": '.PHP_EOL.
+                '- Source: '.$test['source'].PHP_EOL.
+                '- Target: '.$test['target']
+            );
+        }
+    }
 }