+++ changelog.md
@@ -1,3 +1,10 @@
+### v2.3.4 - FileHelper improvements
+- PathInfo: `isWithinFolder()` now correctly handles `../` in paths.
+
+### v2.3.3 - FileHelper improvements
+- PathInfo: Added `getSize()` to get the size in bytes.
+- PathInfo: Added `isWithinFolder()` to compare paths.
+
 ### v2.3.2 - Luma fix
 - RGBAColor: Better Luma calculation, based on an actual Luma formula.
 - RGBAColor: Added `getLumaPercent()`.